Make Google Traffic Raster Based on Grid of Coordinates

Make a raster from Google traffic data, where each pixel has one of four values indicating traffic volume (no traffic, light, moderate, and heavy).

Usage

gt_make_raster_from_grid(
  grid_param_df,
  google_key,
  webshot_delay = NULL,
  return_list_of_tiles = F,
  print_progress = T
)

Arguments

grid_param_df

Grid parameter dataframe produced from gt_make_grid

google_key

Google API key

webshot_delay

How long to wait for google traffic layer to render. Larger height/widths require longer delay times. If NULL, the following delay time (in seconds) is used: delay = max(height,width)/200.

return_list_of_tiles

Instead of merging traffic tiles together into one large tile, return a list of tiles

print_progress

Whether to print function progress

Value

Returns a georeferenced raster file. The file can contain the following values: 1 = no traffic; 2 = light traffic; 3 = moderate traffic; 4 = heavy traffic.

Examples

if (FALSE) {
## Grab polygon of Manhattan
us_sp <- raster::getData('GADM', country='USA', level=2)
ny_sp <- us_sp[us_sp$NAME_2 %in% "New York",]

## Make Grid
grid_df <- gt_make_grid(polygon = ny_sp,
                       height   = 2000,
                       width    = 2000,
                       zoom     = 16)

## Make raster from grid                        
r <- gt_make_raster_from_grid(grid_param_df = grid_clean_df,
                              google_key    = "GOOGLE-KEY-HERE")
}